+++ kubejs/server_scripts/create/recipes.js
@@ -1098,6 +1098,11 @@ const registerCreateRecipes = (event) => {
         
     //#region Покраска ручек от люка
 
+    event.recipes.tfc.barrel_sealed(1000)
+        .inputs('#tfg:colored_valve_handles', Fluid.of(`tfc:lye`, 25))
+        .outputItem(`create:copper_valve_handle`)
+        .id(`barrel/create/valve_handle_decolor`)
+
     event.recipes.gtceu.chemical_bath(`create/valve_handle_decolor`)             
         .itemInputs('#tfg:colored_valve_handles')
         .inputFluids(Fluid.of(`gtceu:chlorine`, 72))
@@ -1106,6 +1111,11 @@ const registerCreateRecipes = (event) => {
         .EUt(4)
 
     global.MINECRAFT_DYE_NAMES.forEach(dye => {
+        event.recipes.tfc.barrel_sealed(1000)
+            .inputs('create:copper_valve_handle', Fluid.of(`tfc:${dye}_dye`, 25))
+            .outputItem(`create:${dye}_valve_handle`)
+            .id(`barrel/create/${dye}_valve_handle`)
+
         event.recipes.gtceu.chemical_bath(`create/${dye}_valve_handle`)             
             .itemInputs('create:copper_valve_handle')
             .inputFluids(Fluid.of(`gtceu:${dye}_dye`, 144))
@@ -1119,18 +1129,32 @@ const registerCreateRecipes = (event) => {
     //#region Покраска тулбоксов
 
     global.MINECRAFT_DYE_NAMES.forEach(dye => {
-        if (dye != 'brown')
+        if (dye != 'brown') {
+            event.remove({ id: `create:create.toolbox.color.block.create.${dye}_toolbox` })
+
+            event.recipes.tfc.barrel_sealed(1000)
+                .inputs('create:brown_toolbox', Fluid.of(`tfc:${dye}_dye`, 50))
+                .outputItem(`create:${dye}_toolbox`)
+                .id(`barrel/create/${dye}_toolbox`)
+    
             event.recipes.gtceu.chemical_bath(`create/${dye}_toolbox`)             
                 .itemInputs('create:brown_toolbox')
                 .inputFluids(Fluid.of(`gtceu:${dye}_dye`, 96))
                 .itemOutputs(`create:${dye}_toolbox`)
                 .duration(200)
                 .EUt(4)
+        }
+        
     })
 
     //#endregion
 
     //#region Покраска сидушек
+
+    event.recipes.tfc.barrel_sealed(1000)
+        .inputs('#tfg:colored_seats', Fluid.of(`tfc:lye`, 25))
+        .outputItem(`create:white_seat`)
+        .id(`barrel/create/seat_decolor`)
 
     event.recipes.gtceu.chemical_bath(`create/seat_decolor`)             
         .itemInputs('#tfg:colored_seats')
@@ -1140,13 +1164,20 @@ const registerCreateRecipes = (event) => {
         .EUt(4)
 
     global.MINECRAFT_DYE_NAMES.forEach(dye => {
-        if (dye != "white")
+        if (dye != "white") {
+            
+            event.recipes.tfc.barrel_sealed(1000)
+                .inputs(`create:white_seat`, Fluid.of(`tfc:${dye}_dye`, 50))
+                .outputItem(`create:${dye}_seat`)
+                .id(`barrel/create/${dye}_seat`)
+
             event.recipes.gtceu.chemical_bath(`create/${dye}_seat`)             
                 .itemInputs(`create:white_seat`)
                 .inputFluids(Fluid.of(`gtceu:${dye}_dye`, 144))
                 .itemOutputs(`create:${dye}_seat`)
                 .duration(300)
                 .EUt(4)
+        }
     })
 
     //#endregion